requirements.txt
boto3==1.24.19
jinja2==3.1.1
markdown==3.3.6
python-dotenv==0.20.0
pyyaml==6.0
watchdog==2.1.7
